dairy_profitability_api/requirements.txt
Django==4.2.20
django-cors-headers==4.9.0
django-mysql==4.19.0
djangorestframework==3.15.2
gunicorn==23.0.0
ipython
mysqlclient==2.2.8
PyJWT==2.9.0
pymysql==1.1.2
python-dotenv==1.2.2
pytz
sqlparse
typing_extensions
tzdata
uuid
pandas
pydruid
